Output Dir Cleanup, Connector API Key Handling
Output Directory is deleted as the aggregations are not tracked as of now.
The data will be re-calculated on every run.
Eventually the idea is to track all the billing data sets, but it is not done at the moment.

Connector API key if kafka.api.key is masked, then all ownership will be defaulted to the kafka cluster ID.